refactor: component FormRegister
