+++ client/src/pages/Register.jsx
@@ -1,69 +1,10 @@
-import { useForm } from 'react-hook-form';
-import { yupResolver } from '@hookform/resolvers/yup';
-import schema from '../validations/schemaValidationRegister';
-
+import FormRegister from "../components/FormRegister";
 function Register() {
-  const { register,
-    handleSubmit: onSubmit,
-    formState: { errors },
-  } = useForm({ resolver: yupResolver(schema) });
 
-  const handleSubmit = (data) => {
-    console.log(data);
-  };
 
   return (
     <div>
-      <form onSubmit={ onSubmit(handleSubmit) }>
-        <label>
-          Nome:
-          <input
-            { ...register('name') }
-          />
-          <span>{errors?.name?.message}</span>
-        </label>
-        <br />
-
-        <label>
-          Email:
-          <input
-            { ...register('email') }
-          />
-          <span>{errors?.email?.message}</span>
-        </label>
-        <br />
-
-        <label>
-          Confirme seu email:
-          <input
-            { ...register('emailConfirm') }
-          />
-          <span>{errors?.emailConfirm?.message}</span>
-        </label>
-        <br />
-
-        <label>
-          Senha:
-          <input
-            type="password"
-            { ...register('password') }
-          />
-          <span>{errors?.password?.message}</span>
-        </label>
-        <br />
-
-        <label>
-          Confirme sua senha:
-          <input
-            type="password"
-            { ...register('passwordConfirm') }
-          />
-          <span>{errors?.passwordConfirm?.message}</span>
-        </label>
-        <br />
-
-        <button type="submit">Entrar</button>
-      </form>
+      <FormRegister />
     </div>
   );
 }
